+++ packages/admin-api-client/CHANGELOG.md
@@ -4,11 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
 ### Added
 ### Changed
 ### Fixed
 ### Removed
+
+## [unreleased]
+
+## [7.0.1] - 2023-05-29
+
+### Changed
+
+- chore(deps): Update dependencies
 
 ## [7.0.0] - 2022-09-14