+++ packages/logger/CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
+## [unreleased]
+
+## [1.1.2] - 2023-05-29
+
+### Changed
+
+- chore(deps): Update dependencies
+
 ## [1.1.1] - 2023-04-09
 
 ### Changed