+++ packages/plugin-admin-api/CHANGELOG.md
@@ -6,11 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [To be deprecated]
 
-## [unreleased]
 ### Added
 ### Changed
 ### Fixed
 ### Removed
+
+## [unreleased]
+
+## [4.0.1] - 2023-05-29
+
+### Changed
+
+- chore(deps): Update dependencies
 
 ## [4.0.0] - 2022-09-14